Module feed::rss

All the structs for rss.

Modules

category
The fields under category can be retrieved by using the methods under Category.

category_builder
The fields can be set for category by using the methods under CategoryBuilder.

channel
The fields under channel can be retrieved by using the methods under Channel.

channel_builder
The fields can be set for channel by using the methods under ChannelBuilder.

cloud
The fields under cloud can be retrieved by using the methods under Cloud.

cloud_builder
The fields can be set for cloud by using the methods under CloudBuilder.

enclosure
The fields under enclosure can be retrieved by using the methods under Enclosure.

enclosure_builder
The fields can be set for enclosure by using the methods under EnclosureBuilder.

guid
The fields under guid can be retrieved by using the methods under Guid.

guid_builder
The fields can be set for guid by using the methods under GuidBuilder.

image
The fields under image can be retrieved by using the methods under Image.

image_builder
The fields under image can be retrieved by using the methods under Image and the fields can be set for image by using the methods under ImageBuilder.

item
The fields under item can be retrieved by using the methods under Item.

item_builder
The fields can be set for item by using the methods under ItemBuilder.

source
The fields under source can be retrieved by using the methods under Source.

source_builder
The fields can be set for source by using the methods under SourceBuilder.

text_input
The fields under text input can be retrieved by using the methods under TextInput.

text_input_builder
The fields can be set for text input by using the methods under TextInputBuilder.

Structs

Category
This Category struct contains all the items that exist for the category field under 'Channel' and Item.

CategoryBuilder
This CategoryBuilder struct creates the Category.

Channel
This Channel struct contains all the items that exist for the feed.

ChannelBuilder
This ChannelBuilder struct creates the Channel.

Cloud
This Cloud struct contains all the items that exist for the cloud field under 'Channel'.

CloudBuilder
This CloudBuilder struct creates the Cloud.

Enclosure
This Enclosure struct contains all the items that exist for the enclosure field under 'Item'.

EnclosureBuilder
This EnclosureBuilder struct creates the Enclosure.

Guid
This Guid struct contains all the items that exist for the guid field under 'Item'.

GuidBuilder
This GuidBuilder struct creates the Guid.

Image
This Image struct contains all the items that exist for the image field under 'Channel'.

ImageBuilder
This ImageBuilder struct creates the Image.

Item
This Item struct contains all the items that exist for the item field under 'Channel'.

ItemBuilder
This ItemBuilder struct creates the Item.

Source
This Source struct contains all the items that exist for the source field under 'Item'.

SourceBuilder
This SourceBuilder struct creates the Source.

TextInput
This TextInput struct contains all the items that exist for the text input field under 'Channel'.

TextInputBuilder
This TextInputBuilder struct creates the TextInput.