Struct feed::rss::ChannelBuilder

pub struct ChannelBuilder {
    // some fields omitted
}

This ChannelBuilder struct creates the Channel.

Methods

impl ChannelBuilder

fn new() -> ChannelBuilder

Construct a new ChannelBuilder and return default values.

Examples

use feed::rss::ChannelBuilder;

let channel_builder = ChannelBuilder::new();

fn title(&mut self, title: &str) -> &mut ChannelBuilder

Set the title that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.title("The Linux Action Show! OGG");

fn link(&mut self, link: &str) -> &mut ChannelBuilder

Set the link that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.link("http://www.jupiterbroadcasting.com");

fn description(&mut self, description: &str) -> &mut ChannelBuilder

Set the description that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let description = "Ogg Vorbis audio versions of The Linux ".to_owned()
+ "Action Show! A show that covers everything geeks care about in the "
+ "computer industry. Get a solid dose of Linux, gadgets, news events "
+ "and much more!";

let mut channel_builder = ChannelBuilder::new();
channel_builder.description(description.as_ref());

fn language(&mut self, language: Option<String>) -> &mut ChannelBuilder

Set the optional language that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.language(Some("en".to_owned()));

fn copyright(&mut self, copyright: Option<String>) -> &mut ChannelBuilder

Set the optional copyright that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let copyright = "Copyright 2002, Spartanburg Herald-Journal".to_owned();

let mut channel_builder = ChannelBuilder::new();
channel_builder.copyright(Some(copyright));

fn managing_editor(&mut self, managing_editor: Option<String>) -> &mut ChannelBuilder

Set the optional managing editor that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let managing_editor =
    "chris@jupiterbroadcasting.com (Chris Fisher)".to_owned();

let mut channel_builder = ChannelBuilder::new();
channel_builder.managing_editor(Some(managing_editor));

fn web_master(&mut self, web_master: Option<String>) -> &mut ChannelBuilder

Set the optional web master that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let web_master =
    "chris@jupiterbroadcasting.com (Chris Fisher)".to_owned();

let mut channel_builder = ChannelBuilder::new();
channel_builder.web_master(Some(web_master));

fn pub_date(&mut self, pub_date: Option<String>) -> &mut ChannelBuilder

Set the optional pub date that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.pub_date(Some("Sun, 13 Mar 2016 20:02:02 -0700".to_owned()));

fn last_build_date(&mut self, last_build_date: Option<String>) -> &mut ChannelBuilder

Set the optional last build date that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.last_build_date(Some("Sun, 13 Mar 2016 20:02:02 -0700".to_owned()));

fn categories(&mut self, categories: Option<Vec<Category>>) -> &mut ChannelBuilder

Set the optional categories that exists under Channel.

Examples

use feed::rss::{ChannelBuilder, CategoryBuilder};

let category = CategoryBuilder::new().finalize();
let categories = vec![category];

let mut channel_builder = ChannelBuilder::new();
channel_builder.categories(Some(categories));

fn generator(&mut self, generator: Option<String>) -> &mut ChannelBuilder

Set the optional generator that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let generator = "Feeder 2.5.12(2294); ".to_owned()
+ "Mac OS X Version 10.9.5 (Build 13F34) "
+ "http://reinventedsoftware.com/feeder/";

let mut channel_builder = ChannelBuilder::new();
channel_builder.generator(Some(generator));

fn docs(&mut self, docs: Option<String>) -> &mut ChannelBuilder

Set the optional docs that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.docs(Some("http://blogs.law.harvard.edu/tech/rss".to_owned()));

fn cloud(&mut self, cloud: Option<Cloud>) -> &mut ChannelBuilder

Set the optional cloud that exists under Channel.

Examples

use feed::rss::{ChannelBuilder, CloudBuilder};

let cloud = CloudBuilder::new().finalize();

let mut channel_builder = ChannelBuilder::new();
channel_builder.cloud(Some(cloud));

fn ttl(&mut self, ttl: Option<i64>) -> &mut ChannelBuilder

Set the optional ttl that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.ttl(Some(60));

fn image(&mut self, image: Option<Image>) -> &mut ChannelBuilder

Set the optional image that exists under Channel.

Examples

use feed::rss::{ChannelBuilder, ImageBuilder};

let image = ImageBuilder::new().finalize();

let mut channel_builder = ChannelBuilder::new();
channel_builder.image(Some(image));

fn rating(&mut self, rating: Option<String>) -> &mut ChannelBuilder

Set the optional rating that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let mut channel_builder = ChannelBuilder::new();
channel_builder.rating(Some("PG-13".to_owned()));

fn text_input(&mut self, text_input: Option<TextInput>) -> &mut ChannelBuilder

Set the optional text input that exists under Channel.

Examples

use feed::rss::{ChannelBuilder, TextInputBuilder};

let text_input = TextInputBuilder::new().finalize();

let mut channel_builder = ChannelBuilder::new();
channel_builder.text_input(Some(text_input));

fn skip_hours(&mut self, skip_hours: Option<Vec<i64>>) -> &mut ChannelBuilder

Set the optional skipdays that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let hours: Vec<i64> = vec![0, 12, 18];

let mut channel_builder = ChannelBuilder::new();
channel_builder.skip_hours(Some(hours));

fn skip_days(&mut self, skip_days: Option<Vec<String>>) -> &mut ChannelBuilder

Set the optional skipdays that exists under Channel.

Examples

use feed::rss::ChannelBuilder;

let days = vec!["Monday".to_owned(), "Tuesday".to_owned()];

let mut channel_builder = ChannelBuilder::new();
channel_builder.skip_days(Some(days));

fn items(&mut self, items: Option<Vec<Item>>) -> &mut ChannelBuilder

Set the optional items that exists under Channel.

Examples

use feed::rss::{ChannelBuilder, ItemBuilder};

let title = "Making Music with Linux | LAS 408".to_owned();

let item = ItemBuilder::new().title(Some(title)).finalize();
let items = vec![item];

let mut channel_builder = ChannelBuilder::new();
channel_builder.items(Some(items));

fn finalize(&self) -> Channel

Construct the Channel from the ChannelBuilder.

Examples

use feed::rss::ChannelBuilder;

let description = "Ogg Vorbis audio versions of The Linux ".to_owned()
+ "Action Show! A show that covers everything geeks care about in the "
+ "computer industry. Get a solid dose of Linux, gadgets, news events "
+ "and much more!";

let channel = ChannelBuilder::new()
        .title("The Linux Action Show! OGG")
        .link("http://www.jupiterbroadcasting.com")
        .description(description.as_ref())
        .language(None)
        .copyright(None)
        .managing_editor(None)
        .web_master(None)
        .pub_date(None)
        .last_build_date(None)
        .categories(None)
        .generator(None)
        .docs(None)
        .cloud(None)
        .ttl(None)
        .image(None)
        .rating(None)
        .text_input(None)
        .skip_hours(None)
        .skip_days(None)
        .items(None)
        .finalize();